% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lam.R
\name{calc_lam}
\alias{calc_lam}
\title{Calculate population growth rate (lambda) from a matrix population model.}
\usage{
calc_lam(M)
}
\arguments{
\item{M}{Matrix population model}
}
\description{
Based on code from Stevens 200x
}
\references{
Stevens, H. 200x.  A primer of ecology in R.
}
