% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_S.R
\name{calc_S}
\alias{calc_S}
\title{Calculate sensitivities and elasticities}
\usage{
calc_S(M)
}
\arguments{
\item{M}{Matrix population model}
}
\description{
Returns a list with 2 items: S and E.
}
\details{
Based on code from Stevens 200x
}
\references{
Stevens, H. 200x.  A primer of ecology in R.
}
