% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_plot_logit.R
\name{gg_plot_logit}
\alias{gg_plot_logit}
\title{Plot output of logistic population growth model using ggplot2}
\usage{
gg_plot_logit(logit.out, axis.txt.sz = 25, axis.ti.sz = 25, title.sz = 30,
  title = "")
}
\arguments{
\item{logit.out}{output of logitistic modeling functions such as logit_discrete()}

\item{axis.txt.sz}{size of labels along x and y axes}

\item{axis.ti.sz}{size of axis titles}

\item{title}{title for for plot.  Default is blank ("")}
}
\description{
Plot output of logistic population growth model using ggplot2
}
