% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_discrete_1step.R
\name{logit_discrete_1step}
\alias{logit_discrete_1step}
\title{Logistic population growth with a discrete time step}
\usage{
logit_discrete_1step(alpha = 0.01, K = NULL, r.d = 1, Nt = 2, ...)
}
\arguments{
\item{alpha}{per capita intraspecific competition impact.}

\item{K}{Carrying capacity; not currently implemented}

\item{r.d}{discrete growth increment}

\item{Nt}{Initial population size before projection}

\item{...}{additional parameters}

\item{time.steps}{number of time steps for model to run}
}
\description{
The function logit_discrete_1step() projects a population one time step forward. (Eg 1 year).
The function logit_discrete_multistep() call logit_discrete_1step() repeated to project the population
forward multiple time steps.
}
\details{
This model can be framed in terms of either alpha, the impact of intraspecific interactions, or K, the
carrying capacity.
}
\examples{
# Project a population of size 2 forward one time step when
## r.d = 1 and alpha = 0.01
logit_discrete_1step(alpha = 0.01,r.d = 1,Nt = 2)

# Project a population of size 20 forward one time step when
## r.d = 1 and K (Carrying capacity) = 15
logit_discrete_1step(K = 15,r.d = 1,Nt = 20, alpha = NULL)

# Model should be defined in terms of alpha OR K, not both.
# If both are set it assumes you meant to set K
logit_discrete_1step(K = 15,r.d = 1,Nt = 20, alpha = 0.01)

# The defaults for the model are set in terms of alpha
# If you set K and don't  dfine anything for alpha
# The function automatically access the deafult value of alpha (0.01)
# Elicity the message as in the previous example
logit_discrete_1step(K = 15,r.d = 1,Nt = 20, alpha = 0.01)

}
\references{
Stevens.  200x.  A primer of ecology in R.
}
