% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_discrete_multistep.R
\name{logit_discrete_multistep}
\alias{logit_discrete_multistep}
\title{Logistic population growth with a discrete time step}
\usage{
logit_discrete_multistep(alpha = 0.01, K = NULL, r.d = 1, Nt = 2,
  time.steps = 15, add.params = TRUE, ...)
}
\arguments{
\item{alpha}{per capita intraspecific competition impact.}

\item{K}{Carrying capacity; not currently implemented}

\item{r.d}{discrete growth increment}

\item{Nt}{Initial population size before projection}

\item{time.steps}{number of time steps for model to run}

\item{add.params}{Logical.  Should underlying parameters be appended to model (r.d, alpha, K)}

\item{...}{additional parameters}
}
\description{
The function logit_discrete_multistep() call logit_discrete_1step() repeated to project the population
forward multiple time steps.
}
\details{
This model can be framed in terms of either alpha, the impact of intraspecific interactions, or K, the
carrying capacity.
}
\references{
Stevens.  200x.  A primer of ecology in R.
}
